# core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.CreateDocumentRequest:
    properties:
      clientExternalReferenceId:
        type: string
      injectedValues:
        additionalProperties: {}
        type: object
      operationType:
        description: CREATE, RENEW, or AMEND (defaults to CREATE)
        type: string
      recipients:
        items:
          $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.CreateRecipientRequest'
        minItems: 1
        type: array
      relatedDocumentId:
        description: Required for RENEW/AMEND
        type: string
      templateVersionId:
        type: string
      title:
        type: string
    required:
    - recipients
    - templateVersionId
    - title
    type: object
...
  github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.DocumentListResponse:
    properties:
      clientExternalReferenceId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      signerProvider:
        type: string
      status:
        type: string
      templateVersionId:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      workspaceId:
        type: string
    type: object
...
  github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Document Assembly System API - Template management and document generation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Doc Engine API
  version: "1.0"
paths:
...
  /api/v1/documents:
    get:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Search by title
        in: query
        name: search
        type: string
      - description: Limit results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.DocumentListResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: List documents
      tags:
      - Documents
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.CreateDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Create and send document
      tags:
      - Documents
  /api/v1/documents/{documentId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.DocumentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Get document
      tags:
      - Documents
  /api/v1/documents/{documentId}/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Cancel document
      tags:
      - Documents
  /api/v1/documents/{documentId}/deprecate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      - description: Deprecation request
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.InternalDeprecateDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.InternalDeprecateDocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Deprecate completed document
      tags:
      - Documents
  /api/v1/documents/{documentId}/events:
    get:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.DocumentEventResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Get document events
      tags:
      - Documents
  /api/v1/documents/{documentId}/invalidate-tokens:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Invalidate all access tokens for a document
      tags:
      - Documents
  /api/v1/documents/{documentId}/pdf:
    get:
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Download signed PDF
      tags:
      - Documents
  /api/v1/documents/{documentId}/recipients:
    get:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.RecipientResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Get document recipients
      tags:
      - Documents
  /api/v1/documents/{documentId}/recipients/{recipientId}/signing-url:
    get:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      - description: Recipient ID
        in: path
        name: recipientId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.SigningURLResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Get signing URL
      tags:
      - Documents
  /api/v1/documents/{documentId}/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.DocumentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Refresh document status
      tags:
      - Documents
...
  /api/v1/documents/statistics:
    get:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_core_usecase_document.DocumentStatistics'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rendis_doc-assembly_core_internal_adapters_primary_http_dto.ErrorResponse'
      summary: Get document statistics
      tags:
      - Documents
...
securityDefinitions:
  AutomationKey:
    description: Automation API key for programmatic access
    in: header
    name: X-Automation-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"